Move tests from roaster-api to roaster-tests
