+++ tests/src/test/resources/org/jboss/forge/roaster/model/util/eclipse_profile.xml
